warranty_service/requirements.txt
Django
djangorestframework
django-filter
psycopg2-binary
django_heroku
markdown
uritemplate
gunicorn
django-health-check
uritemplate
pyyaml
